+++ app/Templates/Default/Assets/js/bbcode.js
@@ -8,3 +8,13 @@ function wrapText(elementID, openTag, closeTag) {
         textArea.value = begin + openTag + selection + closeTag + end;
     }
 }
+function wrapText(elementID, openTag, closeTag) {
+    var textArea = document.getElementById('status_content');
+
+    if (typeof(textArea.selectionStart) != "undefined") {
+        var begin = textArea.value.substr(0, textArea.selectionStart);
+        var selection = textArea.value.substr(textArea.selectionStart, textArea.selectionEnd - textArea.selectionStart);
+        var end = textArea.value.substr(textArea.selectionEnd);
+        textArea.value = begin + openTag + selection + closeTag + end;
+    }
+}